Fixing panics in release mode

+++ src/rf_channel.rs
@@ -303,7 +303,7 @@ impl Devices {
     /// # Returns
     /// An option containing the devices if they were discovered on the bus. If any device did not
     /// properly enumerate, the option will be empty.
-    fn new(manager: &'static I2cBusManager, delay: &mut impl DelayUs<u8>) -> Option<Self> {
+    fn new(manager: &'static I2cBusManager, delay: &mut impl DelayUs<u16>) -> Option<Self> {
         // The ADS7924 and DAC7571 are present on the booster mainboard, so instantiation
         // and communication should never fail.
         let mut dac7571 = Dac7571::default(manager.acquire_i2c());
@@ -441,7 +441,7 @@ impl RfChannel {
     pub fn new(
         manager: &'static I2cBusManager,
         control_pins: ChannelPins,
-        delay: &mut impl DelayUs<u8>,
+        delay: &mut impl DelayUs<u16>,
     ) -> Option<Self> {
         // Attempt to instantiate the I2C devices on the channel.
         match Devices::new(manager, delay) {
